+++ renovate.json
@@ -3,5 +3,6 @@
   "extends": [
     "config:recommended"
   ],
-  "automerge": true
+  "automerge": true,
+  "ignoreTests": true
 }
