+++ src/main/java/com/renotest/rewrite/domain/BookList.java
@@ -1,0 +1,16 @@
+package com.renotest.rewrite.domain;
+
+import java.util.List;
+
+public class BookList {
+
+	private List<Book> data;
+
+	public List<Book> getData() {
+		return data;
+	}
+
+	public void setData(List<Book> data) {
+		this.data = data;
+	}
+}
