+++ src/main/java/com/renotest/rewrite/domain/ObjectWithId.java
@@ -1,0 +1,20 @@
+package com.renotest.rewrite.domain;
+
+public class ObjectWithId {
+    protected long id;
+
+    public ObjectWithId() {
+    }
+
+    public ObjectWithId(long id) {
+        this.id = id;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+}
